requirements.txt
PyYAML
mistralai
streamlit
langchain
pyzotero
PyMuPDF
